requirements.txt
# 衣服分类模型核心依赖
# 适配Python 3.11 + CUDA 12.4

# 核心深度学习框架
torch>=2.1.0,<2.3.0
torchvision>=0.16.0,<0.18.0
pytorch-lightning>=2.1.0,<2.3.0

# 模型库
timm>=0.9.12

# 数据处理
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0

# 配置文件处理
PyYAML>=6.0.1

# 进度条
tqdm>=4.66.0

# API服务
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# 可选依赖（用于数据分析）
matplotlib>=3.7.0  # 可视化
scikit-learn>=1.3.0  # 机器学习工具
albumentations>=1.3.1  # 数据增强

# 系统监控
psutil>=5.9.0

# 生产环境服务器
gunicorn>=21.2.0
